Fixed incorrect usage of FAPI2_RC_PLAT_ERR_SEE_DATA
In case of functions returning ReturnCode, if any of the operation
fails, then FAPI2_RC_PLAT_ERR_SEE_DATA is returned to the caller.
This is incorrect. To fix this issue, a new FAPI RC has been added
 in the error XML. In case of any failure, an object of FAPI RC will
be created. This object will include the actual RC returned by the
called function.

Note: This problem has not been addressed in the below functions as
these functions are mostly called in the hardware procedure and their
usage needs to be fixed and or revisit of the recoverable error log is
required.
  - ReturnCodes sbefifo_hwp_data_istream::get(hwp_data_unit& o_data)
  - ReturnCodes sbefifo_hwp_data_ostream::put(hwp_data_unit i_data)

JIRA: PFSBE-251
Change-Id: I7c77f179eff91573bcb18f735e6350b0d9799952
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/sbe/+/166016
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Rajees P P <rajerpp1@in.ibm.com>