PM HWP: Level2 version of PM elog HWP
Commit implements level 2 version of pst_pm_elog HWP.
It examines various flag register and determines if an
error log is pending. If an error log exists, it is
uploaded to a specific region of HOMER. Once an error
log is uploaded, a hostboot error log is created and
hcode error log is added as a section to that.

EWM: 27942
Change-Id: I542a25e7eb163d469cde5f8587f2f45c00fa66c5
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/ekb-src/+/168165
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: RANGANATHPRASAD G. BRAHMASAMUDRA <prasadbgr@in.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: DAVID DU <daviddu@us.ibm.com>
Reviewed-by: Greg Still <stillgs@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/sbe/+/168200
Reviewed-by: Team <hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>