Updated the chipTypes property of sbeAttrSync tag
The list of chipTypes must be separated by comma for the attribute
tool to work. For the attribute ATTR_POZ_BIST_TIMER, the chipTypes
has been specified as poz, which has not been recognised by attribute
tool. Hence, this must be corrected. As this attribute is not present in
Odyssey SBE filter XML, there is no need for Hostboot/Cronus to send this
attribute for update to SBE. ie. the value of toSBE should not be 1 for
Odyssey. So, I removed ody from the chipTypes.

Change-Id: Id767fb1be96c98a6cf3821300f844ca779da4812
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/ekb-src/+/168244
Reviewed-by: Kahn C Evans <kahnevan@us.ibm.com>
Reviewed-by: Josh Chica <josh.chica@ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/sbe/+/168250
Reviewed-by: Team <hostboot@us.ibm.com>