Added a method to validate sbeAttrSync
If an attribute is not consumed by SBE (attribute not present in filter
XML), then that attribute should not be tagged for syncing from Hostboot/
Cronus ie. toSBE="1". This is to avoid SBE returning attribute not found
error to the caller of the attribute update chip-op. If there are so many
such attributes, then it would increase the size of the attribute update
blob which will result in more time taken for attribute sync istep.

Change-Id: I22c82c017a0c809757d4fff1df6b1534ba792e9c
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/sbe/+/168282
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Anuj Tiwari <anuj.tiwari2@ibm.com>
Reviewed-by: Rajees P P <rajerpp1@in.ibm.com>