Restore trailing newlines on some files
The files in EKB have trailing newlines, the files in SBE do not - looks like
auto mirror conveniently lobbed them off. This is no problem until an incoming
commit touches the end of these files, at which point the incoming patch will
fail to apply on the SBE side.

Prepare for an upcoming patch by restoring the trailing newlines for the
affected files only.

Change-Id: I815ad278d3bbc9aded9e6a9cbabc539578961c7b