Added RC_POZ_STREAM_GET_FAILED
sbefifo_hwp_data_istream::get() method has been modified to return
ReturnCode to the caller. Hence, in case of failure, FAPI RC must be
created and returned to the caller. This RC has been added to create
FFDC if the caller tries to read when words left is zero or data
pointer is NULL.

Change-Id: Ic490e22c11b80ead4793ca202e6e94bfaafd4172
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/ekb-src/+/170460
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Mayur Bansal <mayur.bansal1@ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/sbe/+/170568
Reviewed-by: Sunil Kumar <skumar8j@in.ibm.com>
Reviewed-by: Team <hostboot@us.ibm.com>