spi: Auto-reset controller on bad status
At the beginning of each transaction, check that the controller status
is clean: everything idle, no recorded errors. If not, perform a controller
reset sequence before starting the transaction and fail if the controller
still is in a bad state after reset.

This obsoletes the --reset option in spitool so remove that.

Change-Id: Icb57649a53e8234b965f2a5de1be6a0d6016fa28
STGD: 611866
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/ekb-src/+/170842
Reviewed-by: Sunil Kumar <skumar8j@in.ibm.com>
Reviewed-by: Sumit Kumar <sumit_kumar@in.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/ekb-src/+/171098
Reviewed-by: Mayur Bansal <mayur.bansal1@ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/sbe/+/171124
Reviewed-by: Team <hostboot@us.ibm.com>