Fixed getSbePlatTargetHandle() to correctly handle DIMM target instance
If a target type has multiple entries(due to different chiplet-id) in the
target map, then the target instance for the second and subsequent entries
were returned as invalid.  Hence, those instances were treated as not present
by the calling function. For example, DIMMs have chiplet-id 0xD and 0XE.
The target map has one entry for the chiplet-id 0xD. The DIMM target with this
chiplet-id has instances 0 and 1. Next to this entry, the target map has an
another entry for the chiplet-id 0XE. This chiplet-id has instances 2 and 3.
While searching for the DIMM target instance 2, invalid instance id was
returned. This has been fixed.

The attribute list chip-op sends attributes for the target instances that
are present. As DIMM target instance 2 was returned as invalid, the attributes
for this instance was not returned to the caller of this chip-op.

EWM: 616824
Change-Id: Ie67e2c328b64185a0224cd54e4e6a993a0e8f000
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/sbe/+/172027
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Rajees P P <rajerpp1@in.ibm.com>
Reviewed-by: Sunil Kumar <skumar8j@in.ibm.com>
Reviewed-by: AMIT J TENDOLKAR <amit.tendolkar@in.ibm.com>