Specify the endianness of SBE trace format
The numerical values in SBE trace are stored in big-endian format.
The python script to convert SBE trace to bmc trace uses sys.byteorder
as the default endianness. As eBMC systems follow little-endian,
the numerical values in SBE trace are interpreted as little-endian.
Hence, the numerical values are printed incorrectly in the SBE trace.
This has been fixed in this commit.

EWM: 628913
Change-Id: I801a33b12212b9952579d4bd2be1309fb09a9e11
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/sbe/+/174508
Reviewed-by: Sunil Kumar <skumar8j@in.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>