Updated HASH_KEY to return last index + 1 for an unknown command class
The command class table is indexed using HASH_KEY. If an invalid command
class is passed, then it returns the last index as a result of fall through
condition. To fix this, HASH_KEY macro has been updated to return the last
index + 1 if an invalid command class is specified.

JUPITR: 1097979
Change-Id: I3363239a99c11d98c71d62083e49a0ae53868acc
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/sbe/+/175373
Reviewed-by: Akhilesh S <akhis023@in.ibm.com>
Reviewed-by: Joachim Fenkes <fenkes@de.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Sunil Kumar <skumar8j@in.ibm.com>
Reviewed-by: Rajees P P <rajerpp1@in.ibm.com>