scan: Reduce long rotate threshold
Greatly reduce the threshold for switching from short rotates (which may
run into PCB timeouts if the target chiplet scans very slowly compared to
the PCB clock) to long rotates (which require polling for completion).

Maintain a larger threshold for the QME because the cores generally run at
a high enough speed.

Change-Id: I40b11d9bd170913c7f761ea4ffbf836e6e8205e3
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/ekb-src/+/175788
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: DAVID DU <daviddu@us.ibm.com>
Reviewed-by: Prem Shanker Jha <premjha2@in.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/sbe/+/175793
Reviewed-by: Team <hostboot@us.ibm.com>