seeprom_pad_byte_write_optimisation
Desription:
Add ability to accept a byte pattern and a length (in double words)
as part of the sbeSpiWriteMsgHdr_t for the SEEPROM byte pattern write
operation. The firmware chipop logic should write the byte pattern to
memory locations starting from the given "start_addr" and continuing
for the number of locations determined by the provided double word length
field.

Change-Id: I4aaeaa7300a0bcacdd29e447269c929bd99ed533
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/sbe/+/177677
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Rajees P P <rajerpp1@in.ibm.com>