poz_putRingBackend: Fix error path
Because of the way we split the embedded archive into a trusted
and untrusted part on the SBE, the archive code would detect
a "corrupted archive" if a trusted file was not found. This got
(rightfully) mapped to PLAT_ERR_SEE_DATA in the FAPI layer, so
to work around this the trusted/untrusted image code had to let
any loadEmbeddedFile error code pass.

Of course this led to _any_ error on file load leading to a pass,
which is not what we want.

To fix this, change the archive code to return FILE_NOT_FOUND in
this exact scenario, so that we can narrow the check in the ring
code again.

Change-Id: I8000b0124d6c3e4a5df2c3320ab9952f31e4f15c
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/ekb-src/+/178304
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Joachim Fenkes <fenkes@de.ibm.com>
Reviewed-by: Andrew Singer <andrew.singer@ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel Kiss <kiss@de.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/sbe/+/178306
Reviewed-by: Team <hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>