Fixed the delay in running DQS tracking HWP
- If the modified value of DQS tracking period is less than the current value and
if the counter has already crossed the modified value
or
- If the DQS tracking period is updated from zero and if the counter has already
crossed the updated value
then run the DQS tracking procedure immediately to avoid delay.ie.Don't wait for
the counter to reach 255, reach to 0, and reach the modified value.

If the DQS tracking period is updated from zero and if the counter has already
crossed the updated value, then run the DQS tracking procedure immediately.

EWM: 621171
Change-Id: Iae7592b92a61e5a648bfd0cbff742db4c3992ac5
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/sbe/+/183708
Reviewed-by: Sunil Kumar <skumar8j@in.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Rajees P P <rajerpp1@in.ibm.com>