Enabled attribute update and list chip-op in Metis SBE
As attribute update and list chip-op were not used before, they were
removed from Metis SBE build to reduce image size.

Now, there is a requirement to save and restore attribute values before
and after concurrent code update respectively. For that, the plan is to
use attribute list chip-op to get a blob of attribute values and then
send that blob as a payload of attribute update chip-op to restore the
attribute values. Hence, the attribute update and list chip-op are
enabled in this commit.

JUP: 1119265
Change-Id: Id01404f0d6e5896f2af5bad0d6b3a16ac84da0b4
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/sbe/+/184278
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: AMIT TENDOLKAR <amit.tendolkar@in.ibm.com>