Modify attribute list chip-op to use stream
The attribute list chip-op calculate the expected buffer
size and requests for buffer of that size to generate the
response. If the buffer space is not available, then the
chip-op fails. To avoid this problem, the attribute list
chip-op has been modified to directly stream the response
to SBE FIFO instead of buffering it and then sending it to
SBE FIFO.

JUP: 1119265
Change-Id: Iaf8a62f5c092430e7df1e313e6da84d57fa30c0a
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/sbe/+/184593
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Rajees P P <rajerpp1@in.ibm.com>
Reviewed-by: AMIT TENDOLKAR <amit.tendolkar@in.ibm.com>