Modify attribute update chip-op to use stream
The existing implementation of attribute update chip-op uses buffer
to store the data and then update the attribute tank from the buffer.
On platform, where enough scratch space is not available, it will fail.
So, the attribute update chip-op implementation has been modified to
directly read from the SBE FIFO and update the attribute tank .

JUP: 1119265
Change-Id: I734632766c948656e84938df0d58f610c9d2c341
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/sbe/+/184893
Reviewed-by: Rajees P P <rajerpp1@in.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: AMIT TENDOLKAR <amit.tendolkar@in.ibm.com>