Fixed attribute update chip-op to flush extra bytes
The attribute update chip-op implementation allocates
buffer based on the payload size. If sufficient buffer
space is not be available, then update chip-op fails.
To handle such cases, the implementation has been modified
to use streams. If there are extra bytes at the end, then
it is treated as unexpected data error. If HWP is
used to generate the blob, then there may be some extra
bytes at the end. In such cases, it is failing. This has
been fixed by flushing the extra bytes instead of treating
them as unexpected data.

EWM: 679885
Change-Id: Ie59c55582c5ebd8fe182d057f6f8e551a5977e99
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/sbe/+/185683
Dev-Ready: RAJESWARAN THILLAIGOVINDAN <rajeswgo@in.ibm.com>
Reviewed-by: AMIT TENDOLKAR <amit.tendolkar@in.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Rajees P P <rajerpp1@in.ibm.com>